Fixed mount issue after reboot. Fixes for: EUCA-9203

+++ worker/__init__.py
@@ -42,7 +42,7 @@ def get_block_devices():
     for filename in os.listdir('/dev'):
         if any(filename.startswith(prefix) for prefix in ('sd', 'xvd', 'vd', 'xd')):
             filename = re.sub('\d', '', filename)
-            if not '/dev/'+filename in ret_list:
+            if not '/dev/' + filename in ret_list:
                 ret_list.append('/dev/' + filename)
     return ret_list
 
@@ -64,14 +64,23 @@ def start_worker():
             log.error('failed to find %s or /mnt' % last_dev)
         else:
             if run_as_sudo('mount | grep /mnt > /dev/null') == 1:
-                if run_as_sudo('mkfs.ext3 %s 2>> /tmp/init.log' % last_dev) != 0 or run_as_sudo(
-                                'mount %s /mnt 2>> /tmp/init.log' % last_dev) != 0:
-                    log.error('failed to format and mount %s ' % last_dev)
+                # try to mount first, there is a chance that system was restarted and
+                # last_dev was formatted and prep before
+                if run_as_sudo('mount %s /mnt 2>> /tmp/init.log' % last_dev) != 0:
+                    if run_as_sudo('mkfs.ext3 %s 2>> /tmp/init.log' % last_dev) != 0 or run_as_sudo(
+                                    'mount %s /mnt 2>> /tmp/init.log' % last_dev) != 0:
+                        log.error('failed to format and mount %s ' % last_dev)
+                    else:
+                        log.info('%s was successfully formatted and mounted to /mnt' % last_dev)
+                        if run_as_sudo('mkdir /mnt/imaging %s 2>> /tmp/init.log') != 0 or run_as_sudo(
+                                'chown imaging-worker:imaging-worker /mnt/imaging 2>> /tmp/init.log') != 0:
+                            log.error('could not create /mnt/imaging')
                 else:
-                    log.info('%s was successfully formatted and mounted to /mnt' % last_dev)
-                    if run_as_sudo('mkdir /mnt/imaging %s 2>> /tmp/init.log') != 0 or run_as_sudo(
-                            'chown imaging-worker:imaging-worker /mnt/imaging 2>> /tmp/init.log') != 0:
-                        log.error('could not create /mnt/imaging')
+                    # make sure that /mnt/imaging exist and re-create it if needed
+                    if subprocess.call('ls -la /mnt/imaging > /dev/null', shell=True) != 0:
+                        if run_as_sudo('mkdir /mnt/imaging %s 2>> /tmp/init.log') != 0 or run_as_sudo(
+                                'chown imaging-worker:imaging-worker /mnt/imaging 2>> /tmp/init.log') != 0:
+                            log.error('could not create /mnt/imaging')
             else:
                 log.info('%s is alredy mounted to /mnt' % last_dev)
     except Exception, err:
